backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
supabase==2.10.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
cryptography==44.0.0
pycryptodome==3.21.0
apscheduler==3.11.0
geojson==3.1.0
numpy==2.2.1
httpx>=0.26.0,<0.28
email-validator==2.2.0
pytest==8.3.4
qrcode[pil]==7.4.2
Pillow>=10.0.0
